+++ taproot/index.js
@@ -1,11 +1,13 @@
 const attemptTaprootClaim = require('./attempt_taproot_claim');
 const taprootClaimTransaction = require('./taproot_claim_transaction');
+const taprootCoopPsbt = require('./taproot_coop_psbt');
 const taprootCoopTransaction = require('./taproot_coop_transaction');
 const taprootRefundTransaction = require('./taproot_refund_transaction');
 
 module.exports = {
   attemptTaprootClaim,
   taprootClaimTransaction,
+  taprootCoopPsbt,
   taprootCoopTransaction,
   taprootRefundTransaction,
 };